внесены правки. добавлены анимации

+++ app/src/main/java/com/example/castomview/statusbar/StatusElement.kt
@@ -2,5 +2,5 @@ package com.example.castomview.statusbar
 
 data class StatusElement(
     val color: Int,
-    val value: Float,
+    val value: Int,
 )
